change in naming convention for paginated collections
All paginated collections will be called paginated_collection so they can easily
be navigated (page to page) and filtered/sorted.

Modification to the views were made to accomodate saving the header of the page
modification when the collection gets refreshed. Useful for preserving the state
of the sorted column images.