split models to allow the app support two variations of the models depending on a setting (which has a default) in the settings file
for the RRP and General deployments, variations exist in the checklists
this commit is an attempt to show how this can be accomplished using
the same app. It relies on the ZAMBIA_DEPLOYMENT setting to determine
which variation of the model to load at run time. Being split into
two separate files, it enables us to make modifications at will on one
without implications on the other.

From here, other sections of the app will check this setting and
reconfigure at run time - including templates.