added a custom gp.recipe.pip egg
A custom gp.recipe.pip egg was required due to a build failure
that was a result of an OSX version that no longer supports builds
for the ppc architecture. Somehow, gp.recipe.pip does try to build
universal binaries and libraries on the OSX platform.

This is not a requirement for this project and in order to prevent
the build failure, it was removed.