Fixed bug where time zone was unspecified in OpenRosa header generation.
When generating the OpenRosa headers for the ODK views, the %Z (time zone name)
specifier is used but since we were using naive datetime objects, the time zone
would never be part of the header. By using `datetime.utcnow()` and localizing
to the UTC time zone, we can ensure it generates a time zone name (UTC).