Changed response status code if form was changed.
According to https://bitbucket.org/javarosa/javarosa/wiki/FormSubmissionAPI,
a status code of 201 means everything was okay, and a status code of 202
implies something might have been wrong with the submission. Changed status
code to reflect that if that form was changed after the user downloaded it.

Also, removed the unused `nature` attribute for the `message` element in the
response header.