Load environment file from settings, not manage.py
We've before now loaded the environment file from the management script.
This is fine if you're using the management script, but if, for instance,
you're running tests, the environment file doesn't get loaded and all
hell breaks loose. For that reason, I'm moving the loading of the
environment into the settings file itself. Please correct if this is
bad.