refactor text responses to simplify error messages and only display the most critical error.
when generating error responses, the default was to generate error
responses for cases where both incorrect question codes and extraneous
text was sent. this has been changed such that it prioritizes the
display of incorrect question codes ignoring any extraneous text in the
same text and only displaying the response for extraneous text if no
invalid question codes were sent.