bug fix: when parsing values for QA variables that have non-true values should also trigger an error.
In quality assurance, we check the values of variables defined if they
have a value that is not None. This prevented the parser from handling
situations where the value could be an empty string. This bug fix
generalizes by checking if the value is non-True before returning the
value fixing this problem.