For the message list view, always ensure the currently selected event is used.
If the currently selected event is over, the current implementation of
`overlapping_events()` does not return it, and as such, messages saved with
the currently selected event will not be displayed/exported. This patch
always adds the currently selected event to the list of events for sourcing
messages to be displayed.