added support for google tag manager fixing #5
this is not enabled by default and the
GOOGLE_TAG_MANAGER_KEY parameter needs to be defined
in the environment before it will be enabled; will
also not be enabled if DEBUG is enabled.

in order to make this work, the content security
policy was modified. depending on how this is
implemented in the future, this is something to look
out for as well.