fix: In QA queries, return extra data as string
For QA queries, extra data fields weren't being
resolved to plain strings, but returned as JSONB
column references. This commit fixes that. It is
not, however, a guard against things like trying
to compare an extra data field (always a string)
to a numeric field. That will always fail.

See #27