fix: Update Celery and redis-py
On Python 3.6.7, attempting to run the worker process results in an error.
Updating the Celery version fixes this, but creates a new problem because
the version of redis-py is not pinned, and redis-py 3.X has issues when
using Redis as a Broker for Celery. To fix this, the versions of both
are now pinned.

https://stackoverflow.com/questions/50444988/celery-attributeerror-async-error
https://github.com/celery/kombu/issues/870
https://github.com/celery/kombu/issues/947
https://github.com/andymccurdy/redis-py/issues/1068