feat: make `playback` command accept arguments on command line
The playback command - prior to this commit - would prompt for
the URL and then the file containing the messages. This commit
makes the URL and message file mandatory command line arguments.

The message file can be specified as '-' (without the quotes),
meaning that the file is to be read from stdin.

see: #270