fix: use correct URL for message playback (#282)
* fix: use correct URL for message playback

the playback function, prior to this, used the specified URL to inject
messages. this would require the user to know the path to the Kannel
view. this commit fixes that.

* feat: swap backports.csv for unicodecsv (playback)

use backports.csv in place of unicodecsv for message playback.
the former is more actively developed, and allows use of the `codecs`
module for I/O