Bugfix/locations import (#687)
* chore: remove unused code

* chore: replace deprecated identifier

the function `encrypt_password` from `flask_security.utils`
is deprecated. this commit replaces that name with `hash_password`,
which is a drop-in replacement that isn't deprecated

* feat: enable admin division GPS data toggle

this commit allows toggling on/off of whether an admin division has GPS
data. replaces #678

see: #641, #678, #682

* feat: first working version of location import

* feat: update location import task

* chore: remove 1-to-1 relationships on `LocationPath`

* fix: restrict the selected locales

this commit fixes a bug in location imports where no locations are
imported when multiple locales are enabled for a deployment.