fix: add separate dashboard filter
this commit addresses nditech/apollo-issues#85
by adding a separate sample filter for the dashboard

nditech/apollo-issues#85 is the result of fixing
nditech/apollo-issues#84. since the sample filter no longer filters
based on locations, but participants,
the source for the dashboard, which are master checklists
(that have no participants) cannot be filtered.

this commit creates a new location-based sample filter
specifically for the dashboard.

see: nditech/apollo-issues#84, nditech/apollo-issues#85