Fix double quotes (#818)
* fix: display missing checklist breakdowns

this commit fixes an issue with data summaries, where breakdowns were not displayed.
the immediate cause is that PostgreSQL wraps each item of the result of the function
`array_agg` in double quotes, which eventually gets into the dataframe columns and
data. thus, when during data summary there is an attempt to group by a location
level (say, for example, LGA), the actual column is '"LGA"' and not 'LGA' as expected.
since grouping by a location level is skipped if it does not exist in the dataframe,
breakdowns are not displayed.

this commit replaces double quotes at the beginning or end of each record in the
data summary dataframe, and replaces all double quotes in the column headers.