feat: run event deletion in the worker
this commit addresses nditech/apollo-issues#168 by moving the deletion
out of the web process and into the worker process.
The deletion is done using SQLAlchemy Core, not the ORM.
Finally, a restriction against deleting an event in use by the
initiating admin user is put in place.

see: nditech/apollo-issues#168