docker services update
now using the postgis image directly since our migrations
also create the extension for the database.

added a healthcheck to the database service so other services
can start after reliably confirming that the database service is
healthy.

added a database migrations service so it doesn't have to be done
within the application server.

changed logging level on both celery and gunicorn to reduce logging
information.