fixes issue with exception generated from empty locale
If a participant does not have a locale defined, responses
to their text message should be generated using the
default. This would normally result in the locale string
being blank; however, `force_locale` does not play nice
with blank locales and will throw an exception.

This fix recognizes this fact and would return a
`nullcontext` to be used instead.